% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{crashes}
\alias{crashes}
\alias{crashes_update}
\title{Crash dataset}
\format{Data frames: crashes, crashes_update}
\source{
\url{https://data.cityofnewyork.us/Public-Safety/NYPD-Motor-Vehicle-Collisions/h9gi-nx95}
}
\usage{
crashes
}
\description{
We have provided a dataset consisting of motor vehicle accidents in the borough of Manhattan.
The original data was downloaded from New York City's Open Data initiative.
}
\keyword{datasets}
