% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{cystic_fibrosis}
\alias{cystic_fibrosis}
\title{Cystic fibrosis dataset}
\format{A data frame}
\usage{
cystic_fibrosis
}
\description{
This dataset contains information on cystic fibrosis across 80 families.
}
\keyword{datasets}
