% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{election2016}
\alias{election2016}
\title{2016 election data}
\format{A tibble data frame}
\source{
\url{https://github.com/jdstorey/asdscourse/blob/master/scripts/election_data.R}
}
\usage{
election2016
}
\description{
For this problem, we have provided county level data for the recent 2016
US presidential election. See url below to see how data was generated.
}
\keyword{datasets}
