% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{endtoxin_study}
\alias{endtoxin_study}
\alias{endotoxin_exprs}
\alias{endotoxin_info}
\title{Endotoxin data set}
\format{Two data frames: endotoxin (expression), endotoxin_info (design)}
\description{

}
