% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ExAC}
\alias{ExAC}
\title{Exome Aggregation Consortium (ExAC)}
\format{A tibble data frame}
\source{
\url{http://exac.broadinstitute.org/}
}
\usage{
ExAC
}
\description{
The Exome Aggregation Consortium (ExAC) is a consortium that seeks to combine data from different human exome sequencing projects.
One publicly available portion of their data is the list of all variant sites they have found along with metadata.
We have prepared a small portion (100,000 sites) of that data consisting of identifying information for each variant
(variants are uniquely identified bya reference SNP ID) along with the functional information for each variant as listed in Ensembl.
The glossary for functional terms is available on the Ensembl website.
}
\keyword{datasets}
