% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{package}
\name{fastats}
\alias{fastats}
\alias{fastats-package}
\title{Foundations of Applied Statistics}
\description{
This package contains important functions and datasets that will be used throughout the course.
}
