% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{hedenfalk}
\alias{hedenfalk}
\alias{hedenfalk_info}
\title{Hedenfalk dataset}
\format{Two data frames: hedenfalk (expression), hedenfalk_info (design)}
\usage{
hedenfalk
}
\description{
The gene expression dataset is generated from a two-channel
microarray experiment with a referencesample. Gene expression in
this experiment is measured from RNA purified from the tumors and reference.
The values of the gene expression data are base-2 log of the gene expression
ratio. Morepositive corresponds to more gene expression
in the tumor sample. A value of 1 means there is twice
as much RNA from the tumor sample when compared to RNA
from the reference sample. Similarly, 2 corresponds to four times as
much, -1 corresponds to one half as much, -2 corresponds to one quarter
as much, etc. Note that the gene_idcolumn consists of integers that serve
as placeholders and do not directly map to human genes. There are 22 tumors
analyzed, and the labels and mutations are documented.
}
\keyword{datasets}
