% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{HGDP}
\alias{HGDP}
\title{Human Genome Diversity Project (HGDP) dataset}
\format{A tibble data frame}
\usage{
HGDP
}
\description{
The Human Genome Diversity Project (HGDP) genotyped 1,043 individuals from
Africa, Europe, the Middle East, South and Central Asia, East Asia, Oceania and the Americas.
There are over 650,000 single nucleotide polymorphisms (SNPs). Note that this
dataset is preprocessed and there are 940 individuals and 550,303 SNPs.
}
\keyword{datasets}
