% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{IMDB}
\alias{IMDB}
\title{Internet Movie Database (IMBD)}
\format{A tibble data frame}
\source{
\url{https://www.imdb.com/}
}
\usage{
IMDB
}
\description{
High budget TV shows have been wildly popular in recent years, e.g. the $120
million cost for The Get Down on Netflix and the $100 million cost
for Westworld on HBO. We've sourced data for TV show episodes from the
Internet Movie Database. We pulled all TV show episodes that have been
rated at least 200 times. In TV, the primary creative forces tend to be the
producers, so we have pulled that information as well.
}
\keyword{datasets}
