% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca}
\alias{pca}
\title{Principal Component Analysis}
\usage{
pca(x, space = c("rows", "columns"), nsv = NULL, center = TRUE,
  scale = FALSE)
}
\arguments{
\item{x}{An n x p matrix.}

\item{space}{Specifies which space to do the projections. Either "row" or "column", default is "row".}

\item{nsv}{Optional. An integer specifying the number of PCs.}

\item{center}{If TRUE then matrix will be mean centered. Default is TRUE.}

\item{scale}{If TRUE then matrix will be normalized. Default is FALSE}
}
\value{
A list of object type "pca" containing:
\item{pc}{A matrix containing the principal components.}
\item{loading}{A matrix containing the loadings.}
\item{pve}{A vector containing the percent variation explained by each PC.}
}
\description{
Implements the Principal Component Analysis (PCA) algorithm
}
\examples{
# import data
X <- matrix(rnorm(100 * 10), ncol = 10)

out_pca <- pca(X)

scree_plot(out_pca)
}
