% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{rna_seq}
\alias{rna_seq}
\title{Simulated RNA-Seq data}
\format{A data frame}
\usage{
rna_seq
}
\description{
The rows are the genes and the columns are the
individuals. The data points are counts corresponding to the number of
times the experimenter observed RNA transcribed from that gene.
}
\keyword{datasets}
