% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{scree_plot}
\alias{scree_plot}
\title{Scree plot for PCA object}
\usage{
scree_plot(x, npcs = length(x$pve))
}
\arguments{
\item{x}{pca object}

\item{npcs}{An integer specifying the number of PCs to plot.}
}
\description{
Scree plot for PCA object
}
