% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{yeast}
\alias{yeast}
\alias{exp.e}
\alias{exp.g}
\alias{exp.pos}
\alias{marker}
\alias{marker.pos}
\title{Smith and Kruglyak dataset}
\format{Four data frames: exp.e (expression), exp.g (genotype), exp.pos, marker, marker.pos}
\description{
The yeast data set contains 2820 genotypes in 109 yeast F1 segregants from a cross
between parental lines BY and RM. They also measured gene expression on 4482 genes in each of these segregants when growing in two different
Carbon sources, glucose and ethanol.
}
