% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch-routine-registration.R
\docType{package}
\name{fastats-package}
\alias{fastats}
\alias{fastats-package}
\title{fastats: What the Package Does (One Line, Title Case)}
\description{
High Performance Statistical Test using Rcpp.
}
\author{
\strong{Maintainer}: Huy Nhat Minh Nguyen \email{huynhatminh.nguyen@connect.qut.edu.au}

}
\keyword{internal}
